[Linux] Split BLE helper into separate class files
