+++ src/platform/Linux/bluez/AdapterIterator.h
@@ -21,7 +21,8 @@
 
 #include <gio/gio.h>
 
-#include "lib/core/CHIPError.h"
+#include <lib/core/CHIPError.h>
+#include <platform/Linux/dbus/bluez/DbusBluez.h>
 
 #include "Types.h"
